+++ CHANGELOG.md
@@ -2,12 +2,28 @@
 
 ## 0.3.1 (unreleased)
 
+### added
+
+- plot function for `MeasurementChain` [[#288]](https://github.com/BAMWelDX/weldx/pull/288)
+
 ### ASDF
 
 - remove the `additionalProperties` restriction
   from `single_pass_weld-1.0.0.schema.yaml` [[#283]](https://github.com/BAMWelDX/weldx/pull/283)
 - allow scalar `integer` value in `anyOf` of `time_series-1.0.0.yaml` to
   fix [#282](https://github.com/BAMWelDX/weldx/pull/282) [[#286]](https://github.com/BAMWelDX/weldx/pull/286)
+- add examples to schema files [[#274]](https://github.com/BAMWelDX/weldx/pull/274)
+
+### changes
+
+- `plot_graph` of the CSM now renders static and time-dependent edges differently
+  [[#291]](https://github.com/BAMWelDX/weldx/pull/291)
+
+### fixes
+
+- A warning is now emitted if a `LocalCoordinateSystem` drops a provided time during construction. This usually happens 
+  if the coordinates and orientation only contain a single data point. 
+  [[#285]](https://github.com/BAMWelDX/weldx/pull/285)
 
 ## 0.3.0 (12.03.2021)